Finally Works for textures ! (Rough version). Understand the X offset for the texture (in fact simply number after coma(with modulo for float fmod), to get the percentage of the X pos in the texture. And the textY with simple percent also with texture height. And getting int color is simply the rgb value of (256 * 256 * 256 + 256 * 256 + 256). By creating a number i of line_size * y + bbp/8 * x. pxl_ptr affection become simple. Correct error segfault when dot is not in windows (so could not renderer). And hit wall for both map and map2. (possibility to make a freefly later). Fast rendering for textures. Could optimize draw_background simply to speed render with background.

+++ libft/inc/libft.h
@@ -92,7 +92,7 @@ t_list				*ft_sort_list(t_list *lst, int (*cmp)(int, int));
 
 float				ft_squared(float n);
 float				ft_pythagore(float a, float b);
-double				ft_degreetorad(int degrees);
+double				ft_degreetorad(double degrees);
 double				ft_radtodegree(double rad);
 char				*ft_rgbtohexa(t_rgbcolor rgbcolor);
 char				*ft_rgbftohexa(t_rgbcolorf rgbcolorf);
